backend/requirements.txt
futu-api
schedule
pandas
flask
flask-cors
requests
supabase
python-dotenv
pydantic>=2.0.0
openai>=1.0.0